Closes #45
It was not necessary for SharedPreferences to be loaded as a background
task. This meant it was possible to guarantee that the LiveData exposed
by `SettingsModel` would always have a value, which could be queried
synchronously at launch of `SettingsFragment` for the values. It was
therefore unnecessary to observe for changes, which resulted in the
feedback loop that caused #45.