Closes #49, #66
Guarantee that SelectedJobModel has a synchronous value (i.e., if you
have called `SelectedJobModel.setJob()` at least once and the most
recent call was with a non-null argument then it is
guaranteed that subsequent calls to
`SelectedJobModel.getSelectedJob.getValue()` will not return `null`.

This obviates the need of the `CalculatorFragment` to observe the
`SelectedJobModel`, greatly simplifying the synchronization code by
making it so that the bindings need only be one-way.