Add Recycler View for BoostBox list.
This does not put the code in place to link up the adapter and the
recycler view. This will require changing the collections holding
BoostBoxes returned by the DAOs to be Lists instead of Sets so that they
have guaranteed ordering.

It also does not include the ability to add a boostbox to a job using
the floating action button.

Closes #30.