Switch Bluetooth to use Bluetooth Recycler View.
JobFragment was modified to use the real BoostBoxFragment instead of
MockBluetooth. BluetoothFragment was previously empty and has been
filled in with functionality, and correspondingly bluetooth_layout.xml
has been updated to match. It was neessary to update Job to return a
List of BoostBoxes rather than a Set so that there is a well-defined
ordering for the RecyclerView to use.